+++ packages/health/lib/src/heath_data_types.dart
@@ -71,6 +71,8 @@ enum HealthDataType {
   DISTANCE_CYCLING,
   FLIGHTS_CLIMBED,
   DISTANCE_DELTA,
+  WALKING_SPEED,
+  SPEED,
   MINDFULNESS,
   WATER,
   SLEEP_ASLEEP,
@@ -108,7 +110,7 @@ enum HealthDataType {
   ELECTROCARDIOGRAM,
 
   // Health Connect
-  TOTAL_CALORIES_BURNED
+  TOTAL_CALORIES_BURNED,
 }
 
 /// Access types for Health Data.
@@ -190,6 +192,7 @@ const List<HealthDataType> dataTypeKeysIOS = [
   HealthDataType.DISTANCE_WALKING_RUNNING,
   HealthDataType.DISTANCE_SWIMMING,
   HealthDataType.DISTANCE_CYCLING,
+  HealthDataType.WALKING_SPEED,
   HealthDataType.MINDFULNESS,
   HealthDataType.SLEEP_ASLEEP,
   HealthDataType.SLEEP_AWAKE,
@@ -235,6 +238,7 @@ const List<HealthDataType> dataTypeKeysAndroid = [
   HealthDataType.STEPS,
   HealthDataType.WEIGHT,
   HealthDataType.DISTANCE_DELTA,
+  HealthDataType.SPEED,
   HealthDataType.SLEEP_ASLEEP,
   HealthDataType.SLEEP_AWAKE_IN_BED,
   HealthDataType.SLEEP_AWAKE,
@@ -326,6 +330,8 @@ const Map<HealthDataType, HealthDataUnit> dataTypeToUnit = {
   HealthDataType.DISTANCE_CYCLING: HealthDataUnit.METER,
   HealthDataType.FLIGHTS_CLIMBED: HealthDataUnit.COUNT,
   HealthDataType.DISTANCE_DELTA: HealthDataUnit.METER,
+  HealthDataType.WALKING_SPEED: HealthDataUnit.METER_PER_SECOND,
+  HealthDataType.SPEED: HealthDataUnit.METER_PER_SECOND,
 
   HealthDataType.WATER: HealthDataUnit.LITER,
   HealthDataType.SLEEP_ASLEEP: HealthDataUnit.MINUTE,
@@ -453,6 +459,7 @@ enum HealthDataUnit {
   BEATS_PER_MINUTE,
   RESPIRATIONS_PER_MINUTE,
   MILLIGRAM_PER_DECILITER,
+  METER_PER_SECOND,
   UNKNOWN_UNIT,
   NO_UNIT,
 }